requirements.txt
# Core Dependencies
python-dotenv==1.0.1
pandas>=2.2.0
numpy>=1.26.0

# LLM and AI Framework
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langgraph>=0.0.20

# Data Processing
duckdb>=0.10.0

# UI Framework
streamlit>=1.31.0
plotly>=5.18.0

# Utilities
python-dateutil>=2.8.2
pydantic>=2.5.0
tiktoken>=0.5.0
